meeting-mind/backend/requirements.txt
fastapi
uvicorn
tinydb
langchain
chromadb
openai
python-dotenv
python-multipart
tiktoken
langchain_openai
langchain-community